requirements.txt
unidecode==1.3.6
GitPython==3.1.30
python-frontmatter==1.0.0
